+++ generate.sh
@@ -12,9 +12,10 @@ set -e
 
 # Run test
 extension/lib/DigitalOrigin/bin/phpunit --group magento-basic
-extension/lib/DigitalOrigin/bin/phpunit --group magento-configure-backoffice
+extension/lib/DigitalOrigin/bin/phpunit --group magento-configure-backoffice-iframe
 extension/lib/DigitalOrigin/bin/phpunit --group magento-product-page
 extension/lib/DigitalOrigin/bin/phpunit --group magento-buy-unregistered
+extension/lib/DigitalOrigin/bin/phpunit --group magento-configure-backoffice-redirect
 extension/lib/DigitalOrigin/bin/phpunit --group magento-register
 extension/lib/DigitalOrigin/bin/phpunit --group magento-fill-data
 extension/lib/DigitalOrigin/bin/phpunit --group magento-buy-registered
